Update medplum deps and move thread and message to pages
