+++ examples/medplum-chat-demo/src/pages/MessagePage.tsx
@@ -3,7 +3,7 @@ import { parseReference, resolveId } from '@medplum/core';
 import { Communication, Patient } from '@medplum/fhirtypes';
 import { PatientSummary, useMedplum } from '@medplum/react';
 import { useEffect, useState } from 'react';
-import { CommunicationDetails } from './CommunicationDetails';
+import { CommunicationDetails } from '../components/CommunicationDetails';
 
 interface MessagePageProps {
   readonly message: Communication;
